Feature/240621 blog (#73)
* markdown: set yzhang.markdown-all-in-one as default formatter

Change the default formatter for markdown files from dbaeumer.vscode-eslint
to yzhang.markdown-all-in-one in the cnjimbo.github.io.code-workspace file to
improve code quality and consistency.

Blogpress navigation and content updates:
- Add new '不专业代码' (Unprofessional Coding) section with a JavaScript
  category, indicating a shift in content focus.
- Remove the extensive '面试' (Interview) section, potentially indicating a
  change in the site's purpose or the author's priorities.
- Hide the 'algorithm' category in the blogpress navigation, suggesting a
  reorganization of content or a temporary hide.

* docs(blogpress): update navigation to reflect content changes

Update the Vitepress configuration to adjust the site navigation. This includes
renaming a menu item from 'javascript' to 'github' and updating the '不专业代码'
section to reflect new content categories. The changes indicate a shift in focus
and a possible reorganization of the site's content.