build(deps): update element-plus dependency from 2.7.5 to 2.7.6
This change updates the element-plus package version in the dependencies to ensure
the latest features and fixes are incorporated into the project. The update
process involves modifying package.json to point to the newer version and
potentially updating any code that relies on deprecated or changed functionality.

Updated element-plus package information can be found at:
- Release notes: https://github.com/element-plus/element-plus/releases
- Changelog: https://github.com/element-plus/element-plus/blob/master/CHANGELOG.md
- Comparison between versions: https://github.com/element-plus/element-plus/compare/v2.7.5...v2.7.6