update blog branch from master (#111) (#113)
* feat(blogpress): add RSS weekly summary to config (#109)

Introduce a new RSS feed configuration option for weekly summaries in the blogpress
Vitepress configuration. This enhancement allows for more frequent updates to be
delivered to subscribers through RSS feeds.

* Ci/actions remove events of pr (#110)

* ci: remove converted_to_draft and auto_merge_enabled triggers in PR workflow

Revised the Pull Request Check and Deploy Preview workflow to eliminate the
'converted_to_draft' and 'auto_merge_enabled' event triggers. This change
ensures that the workflow is triggered by relevant PR events and avoids
unnecessary builds when draft PRs are converted or auto-merge status changes.

* ci: modify concurrency configuration in PR deploy preview workflow

Change the concurrency group and cancel-in-progress settings in the
Pull Request Check and Deploy Preview workflow. This update switches
the concurrency group to a dynamic one based on the GitHub event and
PR number or ref name, and enables canceling in-progress jobs to
improve efficiency.

* ci: remove commented-out 'delete' branch from PR deploy workflow

Remove the commented-out 'delete' branch specification from the
Pull Request Check and Deploy Preview workflow to clarify the
intended branches for the workflow and avoid potential confusion.

* ci: update concurrency configuration and remove push events in PR workflow (#112)

* ci: revise Pull Request Check and Deploy Preview workflow to improve efficiency
* ci: remove commented-out 'delete' branch to clarify workflow intent
* ci: eliminate 'converted_to_draft' and 'auto_merge_enabled' triggers