feat(blogpress): use FULL_URL for base URL configuration (#114)
* feat(blogpress): use FULL_URL for base URL configuration

Replace PAGES_BASE with FULL_URL in the Vitepress configuration to
provide more flexibility in specifying the base URL. This change
enables the use of a full URL including the protocol and domain,
where previously only the domain was supported.

* fix(blogpress): correct RSS feed image paths in config

Resolve broken RSS feed image URLs by updating the Vitepress configuration
to use correct relative paths. This ensures that the RSS feeds for both
weekly summaries and the site logo are accessible and display properly.

* fix(blogpress): correct avatar URLs in config for RSS feeds

Resolve broken RSS feed avatar URLs by updating the Vitepress configuration
to use correct relative paths with the ${relativeUrl} variable. This ensures
that the RSS feeds for weekly summaries and the site logo display properly.

Additionally, remove the transformation logic for avatar URLs containing
a specific domain, as this is no longer required with the new path handling.