feat(blogpress): update from upstream
Correct image paths in various markdown files to use ./2022/ or ./sw/ prefixes,
and address typos and inconsistencies in the content. Additionally, exclude the
'template' folder from tsconfig.json in the create-theme package and set Vue.volar
as the default formatter for Vue files in the workspace configuration. Lastly,
update eslint config and VS Code extension recommendations for improved maintainability
and functionality.