chore(dependabot): streamline install command and update action branch
Simplify the installation command in dependabot.yml by removing the conditional
check for the existence of pnpm-lock.yaml. The update-command and
on-changes-command have been updated to use quotes for consistency and to
ensure the commands are executed as intended. Additionally, the pr-branch has
been updated to "master" to align with the current default branch name.